#123: Fix 39 numeric constraint mismatches in flow tool (vibe-kanban f67732b1)
See GitHub issue #123 for full details.

This task involves fixing 39 numeric constraint mismatches in the flow tool.

+++ src/schemas/filters.ts
@@ -2,8 +2,8 @@ import { z } from "zod"
 
 /** Premium filter (min/max) */
 export const premiumFilterSchema = z.object({
-  min_premium: z.number().nonnegative("Premium cannot be negative").default(0).describe("The minimum premium on the alert or trade").optional(),
-  max_premium: z.number().nonnegative("Premium cannot be negative").describe("The maximum premium on the alert or trade").optional(),
+  min_premium: z.number().int().nonnegative("Premium cannot be negative").default(0).describe("The minimum premium on the alert or trade").optional(),
+  max_premium: z.number().int().nonnegative("Premium cannot be negative").describe("The maximum premium on the alert or trade").optional(),
 })
 
 /** Size filter (min/max) */
